inlining AffiliateCredit component within AffiliateCredits
